[Refactor][Manifest] isolate legacy roofline evaluator into separate module
Extract all roofline-evaluator code out of tileops/manifest.py into a
sibling module tileops/manifest_legacy_roofline.py. Re-export the
public symbols (eval_roofline, resolve_roofline_vars, has_roofline_vars,
_safe_eval) from tileops.manifest so callers are not disturbed.

The target design (docs/roofline.md §4.4.6) has no manifest-level
roofline evaluator; the surface extracted here is transitional and
tracked for removal. Isolating it now makes the migration boundary
visible in the codebase, shrinks the eventual deletion to an rm, and
lets grep for manifest_legacy_roofline act as a progress indicator
for consumers still on the old path.

Manifest loading (_load_manifest, load_workloads) stays in manifest.py.
The legacy module accesses _load_manifest via the core module object
(_manifest_core._load_manifest()) so test monkeypatches of
tileops.manifest._load_manifest remain effective.

No behavior change: 166 roofline/manifest/workload/validator tests
pass unchanged.

Closes #1023
